requirements.txt
Flask==2.3.3
requests==2.31.0
google-generativeai==0.3.2
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7
